Optimization: Cache avoidance grids per map
Currently, ApplyAvoidGrid() fetches the avoidance grid map component
for the map of the pawn that the grid is being retrieved for on each
invocation. However, Map::GetComponent() has a noticeable cost due to
it having to iterate through all map components and perform a type check
before finding the requested component, which can add up as ApplyAvoidGrid()
may be called thousands of times per tick given enough pawns.

On a map with 17 pawns and ~100 animals, a 90-second `perf` CPU profile indicates
that approximately 2.1% of CPU time is spent in ApplyAvoidGrid(), almost all
of it inside the GetComponent() call. This is roughly similar to the CPU time
spent in the needs or immunity ticks for all pawns on the map.

As a fix, introduce a per-map cache for the avoidance grid to allow for fast
retrieval of the component without having to go through GetComponent();
this is akin to the optimization that Performance Optimizer employs and allows
for users of this mod to benefit from improved performance when running larger
bases (or simply larger ranches) even if they do not happen to have that mod
installed. A subsequent CPU profile on the same map with the adjusted code
only shows ~0.03% of CPU time spent in ApplyAvoidGrid().

Corresponding `perf report` outputs are attached to the PR.

NB.: I have opted to switch the project file to a new-style SDK project,
since the existing project file was not readily buildable on non-Windows
systems (or a system that did not have RimWorld installed in the particular
location referenced by the project file). The modified project file allows
for building the project on any operating system supported by .NET Core
and does not rely on a particular RimWorld installation path.